+++ route4me-csharp-sdk/Route4MeSDKTest/Examples/ExamplesInfrastructure.cs
@@ -37,7 +37,7 @@ namespace Route4MeSDK.Examples
         public List<string> configKeysToRemove = new List<string>();
         public List<string> CustomNoteTypesToRemove = new List<string>();
         public List<string> OrdersToRemove = new List<string>();
-        public List<int> OrderCustomFieldsToRemove = new List<int>();
+        public List<long> OrderCustomFieldsToRemove = new List<long>();
 
         Order lastCreatedOrder;
 
@@ -1213,7 +1213,7 @@ namespace Route4MeSDK.Examples
 
             foreach (string customNoteType in CustomNoteTypesToRemove)
             {
-                int? customNoteTypeId = allCustomNoteTypes
+                long? customNoteTypeId = allCustomNoteTypes
                     .Where(x => x.NoteCustomType == customNoteType)
                     .FirstOrDefault()
                     ?.NoteCustomTypeID ?? -1;
@@ -1251,7 +1251,7 @@ namespace Route4MeSDK.Examples
             }
         }
 
-        private int? GetCustomNoteIdByName(string customNoteTypeName)
+        private long? GetCustomNoteIdByName(string customNoteTypeName)
         {
             var route4Me = new Route4MeManager(ActualApiKey);
 
@@ -1264,7 +1264,7 @@ namespace Route4MeSDK.Examples
 
             if (allCustomNoteTypes == null || allCustomNoteTypes.Count < 1) return null;
 
-            int? customNoteTypeId = allCustomNoteTypes
+            long? customNoteTypeId = allCustomNoteTypes
                     .Where(x => x.NoteCustomType == customNoteTypeName)
                     .FirstOrDefault()
                     ?.NoteCustomTypeID ?? null;
