+++ route4me-csharp-sdk/Route4MeSDKTest/Examples/Routes/UpdateRouteCustomData.cs
@@ -14,7 +14,7 @@ namespace Route4MeSDK.Examples
         /// <param name="routeDestionationId">Route destination ID</param>
         /// <param name="CustomData">Custom data</param>
         public void UpdateRouteCustomData(string routeId = null,
-                                          int? routeDestionationId = null,
+                                          long? routeDestionationId = null,
                                           Dictionary<string, string> CustomData = null)
         {
             // Create the manager with the api key
